Remove testing functionality no longer used
