+++ temba/channels/tests.py
@@ -600,7 +600,7 @@ class ChannelTest(TembaTest):
 
         # test that editors have the channel of the the org the are using
         other_user = self.create_user("Other")
-        self.create_secondary_org()
+        self.setUpSecondaryOrg()
         self.org2.administrators.add(other_user)
         self.org.editors.add(other_user)
         self.assertFalse(self.org2.channels.all())
@@ -1308,7 +1308,7 @@ class ChannelTest(TembaTest):
         )
 
         # create channel in another org
-        self.create_secondary_org()
+        self.setUpSecondaryOrg()
         Channel.create(self.org2, self.admin2, "RW", "A", "", "+250788382382")
 
         # can claim it with this number, and because it's a fully qualified RW number, doesn't matter that channel is US
@@ -2386,7 +2386,7 @@ class ChannelCountTest(TembaTest):
 class ChannelLogTest(TembaTest):
     def test_channellog_views(self):
         contact = self.create_contact("Fred Jones", "+12067799191")
-        self.create_secondary_org(100_000)
+        self.setUpSecondaryOrg(100_000)
 
         # create unrelated incoming message
         Msg.create_incoming(self.channel, "tel:+12067799191", "incoming msg", contact=contact)
