+++ temba/orgs/tests.py
@@ -1228,7 +1228,7 @@ class OrgTest(TembaTest):
         choose_url = reverse("orgs.org_choose")
 
         # have a second org
-        self.create_secondary_org()
+        self.setUpSecondaryOrg()
         self.login(self.admin)
 
         response = self.client.get(reverse("orgs.org_home"))
