+++ temba/tests/__init__.py
@@ -1,1 +1,2 @@
 from .base import *  # noqa
+from .es import *  # noqa
